requirements.txt
# macOS Vision Agent Dependencies
# 基于MLX-VLM、Hammerspoon、CrewAI和PyAutoGUI的macOS视觉智能体系统所需依赖

# 核心框架
crewai==0.5.0

# MLX相关 - Apple Silicon机器学习框架
mlx>=0.0.8
mlx-vlm>=0.0.4

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# GUI自动化
pyautogui>=0.9.54
pynput>=1.7.6

# 网络和API
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.0

# 数据处理
pandas>=1.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日志和配置
loguru>=0.7.0
python-dotenv>=1.0.0
PyYAML>=6.0

# 异步支持
aiofiles>=23.0.0

# 开发工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# 系统集成
psutil>=5.9.0
watchdog>=3.0.0
click>=8.0.0